Extended testing and compatibility
The testing routine now uses the method that is passed to it rather than the hard coded 'evalHand'.
The message and testingsuite lists have been updated to include the comparing hands and made compatible to pass the arguments to the function.
Functions returning a tuple are now being assigned to the proper amount of variables.