Bugfix in compareHands function
A bug caused the test methods to fail after even hands had been evaluated. When the cards are checked for a better kicker cards are removed. Since the list had not been copied but only referenced the cards had been deleted in-place. Thus all functions checking the Four Of A Kind hand had only 3 cards and failed if used after that specific check.